web-animations/animation-timeline/animation-timeline-play-test.html

<!DOCTYPE html>
<html>
<head>
<title>Animation timeline test</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Returns the time value for this timeline or null if this timeline is not started">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var animationTarget = doc.createElement('div');
    
    var animation = newAnimation(animationTarget);
    var player = document.timeline.play(animation);

    assert_not_equals(player, null, 'Player created by play() method cannot be null');
}, 'Player created by play() method cannot be null');

test(function() {
    var doc = newHTMLDocument();
    var animationTarget = doc.createElement('div');
    
    var animation = newAnimation(animationTarget);
    var player = document.timeline.play(animation);

    assert_equals(player.timeline, document.timeline,
        'The timeline attribute of the newly-created AnimationPlayer object ' +
        'should be set to this timeline object');
}, 'The timeline attribute of the newly-created AnimationPlayer object ' +
    'should be set to this timeline object');

</script>
</body>
</html>
